+++ CHANGELOG.md
@@ -7,10 +7,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `glyphtune.arrays` module.
+
+### Changed
+
+- `Square` objects no longer subclass `Pulse` and now subclass `PeriodicWave`.
+- Moved `FloatArray` from `glyphtune` module to `glyphtune.arrays`.
+- `PeriodicWave` now properly overrides `__eq__` and `__repr__`.
+- Periodic wave representations now explicitly state `phase` keyword.
+- `OperationWaveform` attribute `_operator_kwargs` is now `__operator_kwargs`. 
+
 ### Removed
 
 - Calculus module (including DerivativeWaveform and IntegralWaveform)
 - Frequency modulation
+
+### Fixed
+
+- Various bugfixes and improvements in equality checks.
 
 ## [0.1.0] 2024-03-03